Add files via upload
We could now load our already pretrained model to make inference and prediction. This is an example of Mnist dataset that I trained on my Colab and converted that model to .bin and model.json which contains the weights and architecture of our model. This is an example of how we could any model by converting into .json file and uploading its .bin file.